backend/requirements.txt
# Dependencies

blinker==1.9.0
click==8.3.2
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.3
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pymongo==4.16.0
python-dotenv==1.2.2
Werkzeug==3.1.8
